+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.common.core
 - ProxyBuilder no longer an interface.
 - TaskScheduler no longer an interface.
+- MessageBundle no longer an interface.
 ### is.codion.common.model
 - FilterModel.Builder.visiblePredicate() renamed visible().
 - FilterModel.visible() and filtered() renamed included() and excluded(), related renaming.